+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Changes and Fixes
 
+- Module [AHAB](./docs/ahab/Ahab.md)
+    - Fixed reset behavior when loading fields
+    - Fixed various minor issues
 - Module [MB](./docs/trial-and-error/Mb.md)
     - Added sorting option "Newest" to *v2 mod* browser
     - Improved context menus for tags (#461)